Moved from using RobotPos to using 254's geometery
These geometery utils offer much more utility than RobotPos ever could
and hence this commit marks the beginning of the transfer to them. This
is a somewhat long process as RobotPos is so deeply rooted in this
project. These utils will eventually allow us to simplify odometery to
the point that the util we have written is obsolete...