Ramsete is for the most part fixed.
This will be the last of these long and vague commits, I promise...
Regarding the software: There is still an annoying bug that causes our
odometery values to be off by approximately 0.2 feet over a 10 foot run.
The source of this bug is incredibly elusive, as even after switching to
Aaron(3863)'s odometery stuff, the same bug persists. Nevertheless the
error is very consistent among runs so it is decided at this point to
allow the bug to persist, perhaps simply adjusting the trajectories
slightly to account for this error.

As for Ramsete, the equations have been debugged and it is now working
fairly precisely and consistently. Framework for actual implementation
is in the process of being implemented and will shortly follow.