Added Support for STOPPED state of Auto Controller
This state is intended to represent when the Thread wherein this
controller is run is halted. It does not affect the workings of the
controller and is immediately changed out of this state when update is
called, as that signifies that the thread is indeed running. This is
useful for outside commands and such to gain an insight about the actual
status of the controller.

Also added the setState Method.