Removed STOPPED state from RamseteUtil
On second thought, this state has nothing to do with the functionality
of the controller and so the role of keeping track on if the controller
is running or not has been relegated to a boolean in the implementation
of the controller, Ramsete.java. Also removed setStatus method from the
util as it is redundant with these new changes.